+++ packages/react-components/react-nav/library/src/components/Hamburger/Hamburger.tsx
@@ -1,20 +1,21 @@
 import * as React from 'react';
-import { useHamburger_unstable } from './useHamburger';
-import { renderButton_unstable } from '@fluentui/react-button';
-import { useHamburgerStyles_unstable } from './useHamburgerStyles.styles';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
+import { renderButton_unstable } from '@fluentui/react-button';
+import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
+
+import { useHamburger_unstable } from './useHamburger';
+import { useHamburgerStyles_unstable } from './useHamburgerStyles.styles';
 import type { HamburgerProps } from './Hamburger.types';
 
 /**
- * Hamburger component - TODO: add more docs
+ * Hamburger component - a button that toggles a menu or navigation drawer.
  */
 export const Hamburger: ForwardRefComponent<HamburgerProps> = React.forwardRef((props, ref) => {
   const state = useHamburger_unstable(props, ref);
 
   useHamburgerStyles_unstable(state);
-  // TODO update types in packages/react-components/react-shared-contexts/src/CustomStyleHooksContext/CustomStyleHooksContext.ts
-  // https://github.com/microsoft/fluentui/blob/master/rfcs/react-components/convergence/custom-styling.md
-  // useCustomStyleHook_unstable('useHamburgerStyles_unstable')(state);
+  useCustomStyleHook_unstable('useHamburgerStyles_unstable')(state);
+
   return renderButton_unstable(state);
 }) as ForwardRefComponent<HamburgerProps>;