+++ packages/react-monaco-editor/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-monaco-editor
 
-This log was last generated on Thu, 12 Jun 2025 07:20:14 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Jun 2025 07:21:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.7.340](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.7.340)
+
+Fri, 13 Jun 2025 07:21:03 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.7.339..@fluentui/react-monaco-editor_v1.7.340)
+
+### Patches
+
+- Bump @fluentui/react-charting to v5.24.4 ([PR #34623](https://github.com/microsoft/fluentui/pull/34623) by beachball)
 
 ## [1.7.339](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.7.339)